Symbolize options passed to config when updating with a hash
When using version 1.1.0 I ran into problems where the application
was staying as the default even though the rails config set it
differently. I traced the problem back to key type differences within
the configs. Within the library we use symbols but yaml is loaded as
strings. I added a test but it is difficult to make fail consistently
as the bad behavior is very difficult to reproduce.