Verify DB connections before each task execution
As part of the request handling process a rails application will verify
that database connections are active and will reconnect if any have
disconnected. This add similar support to rcelery tasks that use the
rails environment. The task runner is patched to verify db connections
before dispatching the task for execution.