Adding support for channel and connection recovery.
Instead of forcing the user of the system to have to define how to
recover from AMQP channel and connection failures I am adding in a
default policy. The policy takes advantage of some AMQP gem features
and hooks. The user must opt-in to using the policy and is not
enabled by default. This should increase the resiliency of the library
in the face of network instability.